+++ app/classes/web/panel_handler.py
@@ -831,6 +831,7 @@ class PanelHandler(BaseHandler):
                 page_data["config-json"] = self.helper.get_categorized_settings(
                     self.helper.get_all_settings()
                 )
+
                 page_data["availables_languages"] = []
                 page_data["all_languages"] = []
                 page_data["all_partitions"] = self.helper.get_all_mounts()