+++ app/frontend/static/assets/js/shared/files.js
@@ -279,7 +279,7 @@ function setup_table_listeners() {
         // Prevent the click from firing if it’s on the context menu button
         if ($(e.target).closest(".context-button").length) return;
         if ($(e.target).closest(".row-select").length) return;
-        if (!$(this).data("can_open")) return;
+        if (!$(this).data("can_open") && !e.altKey) return; // Allow opening override with alt key + click
         if ($(this).children(".column-1").hasClass("editing")) return;
         window.open(`/panel/edit_file?server_id=${serverId}&file=${encodeURI($(this).attr("data-path"))}`, "_blank")
     });