+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 TBD
 ### Bug fixes
 - Fix crash on startup due to big bucket cache refresh ([Merge Request](https://gitlab.com/crafty-controller/crafty-4/-/merge_requests/941))
+- Fix issue where files would incorrectly be classified as not able to be opened by missing mimetypes ([Merge Request](https://gitlab.com/crafty-controller/crafty-4/-/merge_requests/944))
 ### Tweaks
 TBD
 ### Lang